requirements.txt
python-dotenv
requests
# textual